Push branches.
I pushed all branches, which are Create, Stats, Study, and Library. However, in the App.jsx file, I didn't call the Library file because I'm not sure where the right place to call it is. After checking, you can call "Library" in the right place.

For example, here's how you can call it:

<div className={styles.library}>
  <Library
    selectedItems={selectedItems}
    imageDirectory="/path/to/layout/images"
  />
</div>